21CSMP67/requirements.txt
requests
google.generativeai
customtkinter
tk
pillow
python-tkdnd
tkinterdnd2


